+++ src/main/java/com/example/tiggle/controller/piggybank/PiggyBankController.java
@@ -1,15 +1,13 @@
-package com.example.tiggle.controller.piggy;
+package com.example.tiggle.controller.piggybank;
 
 import com.example.tiggle.dto.common.ApiResponse;
-import com.example.tiggle.dto.piggy.request.CreatePiggyBankRequest;
-import com.example.tiggle.dto.piggy.request.PiggyEntriesPageRequest;
-import com.example.tiggle.dto.piggy.request.UpdatePiggySettingsRequest;
-import com.example.tiggle.dto.piggy.response.PiggyBankResponse;
-import com.example.tiggle.dto.piggy.response.PiggyEntriesPageResponse;
-import com.example.tiggle.dto.piggy.response.PiggySummaryResponse;
-import com.example.tiggle.service.piggy.PiggyCreationService;
-import com.example.tiggle.service.piggy.PiggyService;
-import com.example.tiggle.service.piggy.PiggySummaryService;
+import com.example.tiggle.dto.piggybank.request.CreatePiggyBankRequest;
+import com.example.tiggle.dto.piggybank.request.PiggyBankEntriesPageRequest;
+import com.example.tiggle.dto.piggybank.request.UpdatePiggyBankSettingsRequest;
+import com.example.tiggle.dto.piggybank.response.PiggyBankResponse;
+import com.example.tiggle.dto.piggybank.response.PiggyBankEntriesPageResponse;
+import com.example.tiggle.dto.piggybank.response.PiggyBankSummaryResponse;
+import com.example.tiggle.service.piggybank.PiggyBankService;
 import com.example.tiggle.util.JwtUtil;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
@@ -19,31 +17,29 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
-@RequestMapping("/api/piggy")
+@RequestMapping("/api/piggybank")
 @RequiredArgsConstructor
 @Tag(name = "저금통 설정/조회", description = "저금통 기본 정보 및 ESG 카테고리 설정")
-public class PiggyController {
+public class PiggyBankController {
 
-    private final PiggyService piggySettingsService;
-    private final PiggySummaryService piggySummaryService;
-    private final PiggyCreationService piggyCreationService;
+    private final PiggyBankService piggyBankService;
 
     @Operation(summary = "내 저금통 조회")
     @GetMapping
     public ResponseEntity<ApiResponse<PiggyBankResponse>> getMyPiggy() {
         Integer userId = JwtUtil.getCurrentUserId();
         // 서비스에서 없으면 ResponseStatusException(404) 던짐 → 전역 핸들러가 그대로 내려줌
-        ApiResponse<PiggyBankResponse> body = piggySettingsService.getMyPiggy(userId).block();
+        ApiResponse<PiggyBankResponse> body = piggyBankService.getMyPiggy(userId).block();
         return ResponseEntity.ok(body);
     }
 
     @Operation(summary = "저금통 설정 수정", description = "name/targetAmount/autoSaving/autoDonation/esgCategoryId(선택) 부분 수정")
     @PatchMapping("/settings")
     public ResponseEntity<ApiResponse<PiggyBankResponse>> updateSettings(
-            @Valid @RequestBody UpdatePiggySettingsRequest request
+            @Valid @RequestBody UpdatePiggyBankSettingsRequest request
     ) {
         Integer userId = JwtUtil.getCurrentUserId();
-        ApiResponse<PiggyBankResponse> body = piggySettingsService.updateSettings(userId, request).block();
+        ApiResponse<PiggyBankResponse> body = piggyBankService.updateSettings(userId, request).block();
         return ResponseEntity.ok(body);
     }
 
@@ -51,7 +47,7 @@ public class PiggyController {
     @PutMapping("/category/{categoryId}")
     public ResponseEntity<ApiResponse<PiggyBankResponse>> setCategory(@PathVariable Long categoryId) {
         Integer userId = JwtUtil.getCurrentUserId();
-        ApiResponse<PiggyBankResponse> body = piggySettingsService.setCategory(userId, categoryId).block();
+        ApiResponse<PiggyBankResponse> body = piggyBankService.setCategory(userId, categoryId).block();
         return ResponseEntity.ok(body);
     }
 
@@ -59,39 +55,39 @@ public class PiggyController {
     @DeleteMapping("/category")
     public ResponseEntity<ApiResponse<PiggyBankResponse>> unsetCategory() {
         Integer userId = JwtUtil.getCurrentUserId();
-        ApiResponse<PiggyBankResponse> body = piggySettingsService.unsetCategory(userId).block();
+        ApiResponse<PiggyBankResponse> body = piggyBankService.unsetCategory(userId).block();
         return ResponseEntity.ok(body);
     }
 
     @Operation(summary = "저금통 정보 조회(Summary)")
     @GetMapping("/summary")
-    public ResponseEntity<ApiResponse<PiggySummaryResponse>> getSummary() {
+    public ResponseEntity<ApiResponse<PiggyBankSummaryResponse>> getSummary() {
         String encryptedUserKey = JwtUtil.getCurrentEncryptedUserKey();
         Integer userId = JwtUtil.getCurrentUserId();
-        ApiResponse<PiggySummaryResponse> body = piggySummaryService.getSummary(encryptedUserKey, userId).block();
+        ApiResponse<PiggyBankSummaryResponse> body = piggyBankService.getSummary(encryptedUserKey, userId).block();
         return ResponseEntity.ok(body);
     }
 
     @Operation(summary = "저금통 계좌 개설")
     @PostMapping
-    public ResponseEntity<ApiResponse<PiggySummaryResponse>> createPiggy(
+    public ResponseEntity<ApiResponse<PiggyBankSummaryResponse>> createPiggy(
             @Valid @RequestBody CreatePiggyBankRequest request
     ) {
         String encryptedUserKey = JwtUtil.getCurrentEncryptedUserKey();
         Integer userId = JwtUtil.getCurrentUserId();
-        ApiResponse<PiggySummaryResponse> body = piggyCreationService.create(encryptedUserKey, userId, request).block();
+        ApiResponse<PiggyBankSummaryResponse> body = piggyBankService.create(encryptedUserKey, userId, request).block();
         return ResponseEntity.ok(body);
     }
 
     @Operation(summary = "저금통 적립 내역 조회(커서 페이지네이션)")
     @PostMapping("/entries")
-    public ResponseEntity<ApiResponse<PiggyEntriesPageResponse>> getEntriesPage(
-            @Valid @RequestBody PiggyEntriesPageRequest request
+    public ResponseEntity<ApiResponse<PiggyBankEntriesPageResponse>> getEntriesPage(
+            @Valid @RequestBody PiggyBankEntriesPageRequest request
     ) {
         String encryptedUserKey = JwtUtil.getCurrentEncryptedUserKey();
         Integer userId = JwtUtil.getCurrentUserId();
-        ApiResponse<PiggyEntriesPageResponse> body =
-                piggySummaryService.getEntriesPage(encryptedUserKey, userId, request).block();
+        ApiResponse<PiggyBankEntriesPageResponse> body =
+                piggyBankService.getEntriesPage(encryptedUserKey, userId, request).block();
         return ResponseEntity.ok(body);
     }
 }
