+++ src/main/java/com/example/tiggle/dto/piggybank/request/UpdatePiggyBankSettingsRequest.java
@@ -1,4 +1,4 @@
-package com.example.tiggle.dto.piggy.request;
+package com.example.tiggle.dto.piggybank.request;
 
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.Size;
@@ -7,7 +7,7 @@ import lombok.Data;
 import java.math.BigDecimal;
 
 @Data
-public class UpdatePiggySettingsRequest {
+public class UpdatePiggyBankSettingsRequest {
     @Size(max = 50)
     private String name;
