+++ src/main/java/com/example/tiggle/dto/piggybank/response/PiggyBankEntriesPageResponse.java
@@ -1,12 +1,12 @@
-package com.example.tiggle.dto.piggy.response;
+package com.example.tiggle.dto.piggybank.response;
 
 import lombok.AllArgsConstructor; import lombok.Data;
 import java.util.List;
 
 @Data
 @AllArgsConstructor
-public class PiggyEntriesPageResponse {
-    private List<PiggyEntryItemDto> items;
+public class PiggyBankEntriesPageResponse {
+    private List<PiggyBankEntryItemDto> items;
     private String nextCursor; // 더 없으면 null
     private int size;
     private boolean hasNext;
