+++ src/main/java/com/example/tiggle/dto/piggybank/response/PiggyBankResponse.java
@@ -1,4 +1,4 @@
-package com.example.tiggle.dto.piggy.response;
+package com.example.tiggle.dto.piggybank.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
