+++ src/main/java/com/example/tiggle/dto/piggybank/response/PiggyBankSummaryResponse.java
@@ -1,4 +1,4 @@
-package com.example.tiggle.dto.piggy.response;
+package com.example.tiggle.dto.piggybank.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -7,7 +7,7 @@ import java.math.BigDecimal;
 
 @Data
 @AllArgsConstructor
-public class PiggySummaryResponse {
+public class PiggyBankSummaryResponse {
     private String name;
     private BigDecimal currentAmount;
     private BigDecimal lastWeekSavedAmount;
